Second refactor for organization and backwards compatibility
