+++ vendorconsent/consent.go
@@ -5,8 +5,8 @@ import (
 	"strings"
 
 	"github.com/prebid/go-gdpr/api"
-	"github.com/prebid/go-gdpr/consent1"
-	"github.com/prebid/go-gdpr/consent2"
+	tcf1 "github.com/prebid/go-gdpr/vendorconsent/tcf1"
+	tcf2 "github.com/prebid/go-gdpr/vendorconsent/tcf2"
 )
 
 // ParseString parses a Raw (unpadded) base64 URL encoded string.
@@ -18,7 +18,17 @@ func ParseString(consent string) (api.VendorConsents, error) {
 	}
 	version := uint8(decoded[0] >> 2)
 	if version == 2 {
-		return consent2.Parse(decoded)
+		return tcf2.Parse(decoded)
 	}
-	return consent1.Parse(decoded)
+	return tcf1.Parse(decoded)
+}
+
+// Backwards compatibility
+
+type VendorConsents interface {
+	api.VendorConsents
+}
+
+func Parse(data []byte) (api.VendorConsents, error) {
+	return tcf1.Parse(data)
 }
