+++ vendorconsent/tcf1/test_utils.go
@@ -1,4 +1,4 @@
-package consent1
+package vendorconsent
 
 import (
 	"encoding/base64"
