+++ vendorconsent/tcf2/bitfield.go
@@ -1,4 +1,4 @@
-package consent1
+package vendorconsent
 
 import (
 	"fmt"
@@ -7,12 +7,12 @@ import (
 func parseBitField(data consentMetadata) (*consentBitField, error) {
 	vendorBitsRequired := data.MaxVendorID()
 
-	// BitFields start at bit 173. This means the last three bits of byte 21 are part of the bitfield.
+	// BitFields start at bit 230. This means the last three bits of byte 28 are part of the bitfield.
 	// In this case "others" will never be used, and we don't risk an index-out-of-bounds by using it.
 	if vendorBitsRequired <= 3 {
 		return &consentBitField{
 			consentMetadata: data,
-			firstThree:      data[21],
+			firstTwo:        data[28],
 			others:          nil,
 		}, nil
 	}
@@ -21,23 +21,23 @@ func parseBitField(data consentMetadata) (*consentBitField, error) {
 	if (vendorBitsRequired-3)%8 > 0 {
 		otherBytesRequired = otherBytesRequired + 1
 	}
-	dataLengthRequired := 22 + otherBytesRequired
+	dataLengthRequired := 28 + otherBytesRequired
 	if uint(len(data)) < uint(dataLengthRequired) {
 		return nil, fmt.Errorf("a BitField for %d vendors requires a consent string of %d bytes. This consent string had %d", vendorBitsRequired, dataLengthRequired, len(data))
 	}
 
 	return &consentBitField{
 		consentMetadata: data,
-		firstThree:      data[21],
-		others:          data[22:],
+		firstTwo:        data[28],
+		others:          data[29:],
 	}, nil
 }
 
 // A BitField has len(MaxVendorID()) entries, with one bit for every vendor in the range.
 type consentBitField struct {
 	consentMetadata
-	firstThree byte
-	others     []byte
+	firstTwo byte
+	others   []byte
 }
 
 func (f *consentBitField) VendorConsent(id uint16) bool {
@@ -46,15 +46,12 @@ func (f *consentBitField) VendorConsent(id uint16) bool {
 	}
 	// Careful here... vendor IDs start at index 1...
 	if id <= 3 {
-		return byteToBool(f.firstThree & (0x08 >> id))
+		return byteToBool(f.firstTwo & (0x04 >> id))
 	}
-	return isSet(f.others, uint(id-4))
+	return isSet(f.others, uint(id-3))
 }
 
 // byteToBool returns false if val is 0, and true otherwise
 func byteToBool(val byte) bool {
-	if val == 0 {
-		return false
-	}
-	return true
+	return val != 0
 }
