+++ vendorconsent/tcf2/bitfield_test.go
@@ -1,4 +1,4 @@
-package consent2
+package vendorconsent
 
 import (
 	"testing"
